Fixed a bug where it crashed if reset was true but the directory did not exist

+++ Unreal99_Linux/Program.cs
@@ -321,7 +321,10 @@ class Program
 
         if (remove)
         {
-            Directory.Delete(files, true);
+            if (files != null && Directory.Exists(files))
+            {
+                Directory.Delete(files, true);
+            }
         }
         else
         {
